Time: 12 ms (45.41%), Space: 17.4 MB (63.85%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -19,11 +19,11 @@ public:
         while(!q.empty()){
             int n = q.size();
             auto p = q.front();
-            long long mini = p.second;
+            int mini = p.second;
             int start,end;
             for(int i=0;i<n;i++){
                 auto p = q.front();
-                long long val = p.second - mini;
+                long val = p.second - mini;
                 auto node = p.first;
                 q.pop();
                 if(node->left) q.push({node->left,val*2 + 1});
